app/models: changed the return value of the MatchPassword function.
Passing an object allocated on heap is dangerous when using gorp. Therefore, we
just pass now the reference of an object allocated on stack. Moreover, this
doesn't do well with the previous return signature. Since the only place that
calls this function just need the user's id, we'll just return the id.